test_: restructure tests wakuext messages
