+++ tests-functional/tests/test_wakuext_default_messaging.py
@@ -6,16 +6,6 @@ from steps.messenger import MessengerSteps
 @pytest.mark.usefixtures("setup_two_unprivileged_nodes")
 class TestDefaultMessaging(MessengerSteps):
 
-    def test_one_to_one_messages(self):
-        responses = self.one_to_one_message(5)
-
-        for response in responses:
-            self.receiver.verify_json_schema(response, method="wakuext_sendOneToOneMessage")
-
-            chat = response["result"]["chats"][0]
-            assert chat["id"] == self.receiver.public_key
-            assert chat["lastMessage"]["displayName"] == self.sender.display_name
-
     def test_add_contact(self):
         self.add_contact(execution_number=1, network_condition=None, privileged=False)
